Move save_results to relevant utils module
