+++ src/gnn_reco/components/utils.py
@@ -1,16 +1,6 @@
 import os
 import pandas as pd
 import sqlite3
-import torch
-
-def save_results(db, tag, results, archive,model):
-    db_name = db.split('/')[-1].split('.')[0]
-    path = archive + '/' + db_name + '/' + tag
-    os.makedirs(path, exist_ok = True)
-    results.to_csv(path + '/results.csv')
-    torch.save(model.cpu(), path + '/' + tag + '.pkl')
-    print('Results saved at: \n %s'%path)
-    return
 
 def check_db_size(db):
     max_size = 5000000
