+++ src/gnn_reco/models/training/utils.py
@@ -1,3 +1,5 @@
+import os
+
 from sklearn.model_selection import train_test_split
 import torch
 from torch_geometric.data.batch import Batch
@@ -22,3 +24,11 @@ def make_train_validation_dataloader(db, selection, pulsemap, batch_size, featur
     validation_dataset = SQLiteDataset(db, pulsemap, features, truth, selection=validation_selection)
     validation_dataloader = torch.utils.data.DataLoader(validation_dataset, **common_kwargs)
     return training_dataloader, validation_dataloader
+
+def save_results(db, tag, results, archive,model):
+    db_name = db.split('/')[-1].split('.')[0]
+    path = archive + '/' + db_name + '/' + tag
+    os.makedirs(path, exist_ok = True)
+    results.to_csv(path + '/results.csv')
+    torch.save(model.cpu(), path + '/' + tag + '.pkl')
+    print('Results saved at: \n %s'%path)
